move code to downloadAsPdf.ts
